Updated .doap file
Adding Alejandro Piñeiro as maintainer, as will start to review
patches from the community, and making releases if needed.

Also removed MAINTAINERS file, as the list of maintainers is already
at the doap file, so it is better to avoid having the same info twice
(and probably out of sync).

Updated AUTHORS file accordingly.