FIX SIGEV fatal error
There was a SIGEV fatal error when running orca meaning
the J2SE-access-bridge script was not identified. This
was caused by some dodgy jni initialization and also a
configuration bug in the keymaps. Resolving the issues
allows the program to init but there is still some
locking up in the g_cond(wait) to resolve.

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=742597