Make sure atk objects exist while processing signals
Since we have to avoid calling methods of applications objects, we can't
afford unreferencing from the jaw thread. This change thus makes it just
queue the para struct to be freed, so it's all processed in the application
when it makes the next accessibility call.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=793543