Fix atk event parameters
glib signals have a determined number of parameters, giving an extra 0 is
not useful, and missing parameters are of course a problem :) The attached
patch fixes the existing mismatches. For bounds_changed, the bogus values
could be problematic. It happens that orca actually ignores this kind of
event, so it will not be bothered. Provide a (even bogus) parameter is
however needed for proper marshalling etc. otherwise undefined behavior
would happen.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=793821