Remove gobjects from GC rather than from jaw_impl_finalize
If a gobject still has some reference on GC, we don't actually drop it,
and thus jaw_impl_finalize won't get called, and thus not removed from
objectTable. On next GC, it might still be there, and we will wrongly unref
it again, leading the reference count mismatch.

Better thus just remove the gobject from objectTable immediately. We
don't actually need to do it from jaw_impl_finalize since as long as
it's in objectTable we have a reference on it so jaw_impl_finalize can't
be called before we drop that reference anyway.

Fixes #33