Do not add/remove references for children
We do not actually need to track references for children to keep the
parent alive, at worse the parent would be indeed dropped only if it's
GC-ed, and thus get_parent returning NULL is right.

This was introduced by 0dd40b694542 ("Fix missing reference addition on
children_changed::add") only because the unref was spurious.